RFC: start in making plasma deps optional
Summary:
make a wrapper for theme and units: one using plasma deps and one
without, with more hardcoded values, eventually what to installlshould be decided
at build time.
To be possible to have a full application done with it, some components should be
cloned from plasma components and plasma stuff removed.
right now the clones are PageRow, Heading, Label.

porting is still not complete, more of an RFC for now

Reviewers: #plasma:_mobile

Subscribers: plasma-devel

Differential Revision: https://phabricator.kde.org/D502