disable gammadjust in Icon
This works around a bug seen on some Android hardware such as Adreno. It
seems that this ends up calling into shaders which screw up the texture,
leading to black squares instead of icons.

We'll need a proper fix for this, either in the drivers, in Qt, or using
a different effect to paint the active state.

For now: sacrifice this functionality for stability :/