Make all GlobalDrawer contents scrollable
With menus longer than few items, the layout of GlobalDrawer gets broken
in landcape mode, where the available height is really small.
put the whole contents in a flickable to be able to use
the content.
this is a behavior similar to similar components in Android
such as the drawer in the GMail app