split ApplicationWindow in AbstractApplicationWindow
it's possible that some kind of applications will not
need/can not use PageRow.
AbstractApplicationWindow provides the basic functionality
such as support for drawers, headers and back button,
but doesn't include the whole PageRow logic.

The ApplicationWindow extends it providing the PAgeRow,
keeping the same api and behavior it had before the split