Polish SwipeListItem
Summary:
Make sure we don't have 2 different code paths for contentItem
initialization as one of them had a different parent.
Set a negative z to the background, to make sure it's below the contents.
Don't let the padding go over the handle, this way the contens will be
able to fit to the space

Test Plan:
Looks the same on kirigamigallery
Now we can make it look okay in discover, as some neon sources get really long.

Reviewers: #kirigami, mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D2827