Make sure that triggering a menu action keeps respecting its action value
Summary:
Kirigami.Action.checked is designed so that the user has to change its value
upon triggered rather than having it change automatically.
Make sure this behavior is respected on the GlobalDrawer

Test Plan:
In discover, when showing a section already, it doesn't think it's
changed when clicking it again.

Reviewers: #kirigami, mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D3764