ItemViewHeader
Summary:
new elements: AbstractItemViewHeader and ItemViewHeader.
An item that can be used as an header for a ListView.
It will play nice with the margin policies of ScrollablePage and can
automatically shrink when the list is scrolled, like the behavior
of list headers in many mobile applications.
It provides some default content: a title and an optional background image

Test Plan: works in gallery, plan to write a new dedicated page

Reviewers: apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D4635