Polish ItemViewHeader, solve binding loops
Summary:
Don't assume "page" in the namespace of ItemViewHeader, use view instead
which we have and symbolizes the same.
Don't change the implicitHeight when contentY is < 0.

Smooth ItemViewHeader

Make page property readonly. Nobody is accessing it and it's private
anyway.
When there's an image, don't wait to have the image to resize the header,
it makes the whole UI change without much sense as it should be fast.
Properly access the page property by name of the object, rather than having
QML go through the object tree. It failed to find the property sometimes.

Test Plan: Tests still pass, I don't get tons of warnings on Discover

Reviewers: #kirigami, leinir, mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D5416