Don't forward keys until the recipient is ready
This avoids numerous crashes when using the keyboard navigation
in Discover during searches. It would seem that, without this,
we occasionally (which during a search with auto-updating results
means regularly) send key events to a component which is not yet
ready to receive them... It would be nice if this were caught by
Qt, but i guess that might be a little expensive to do generically.

nb: i would like this to be a typesafe comparison, but... yay qml ;)

Differential Revision: https://phabricator.kde.org/D5408

BUG: 378339