Honour layer position for isCurrentPage
If using layers, isCurrentPage would without this patch
potentially be conceptually true for more than one page
at a time, but in fact only true for the current page of
the PageRow, and not the page in the layer currently on top.
This seems awkward and seems perhaps unintended, so, this
patch makes isCurrentPage honour layers as well as the
PageRow itself.

Differentia Revision: https://phabricator.kde.org/D7466