Make passiveNotifications last longer, so users can actually read them
Summary:
Right now passiveNotifications last for 4.5 seconds by default, including the time taken up by their appearance and disappearance animations. In practice, the notifications are readable for abour 3.5-4 seconds, max. This is far too short a time for the user to notice that the notification appeared, move their eyeballs to it, start reading it, and finish reading it.

When a notification disappears before it's been fully read, or worse, before the user even started reading it, extreme feelings of user discomfort are generated. Users start to think, "did I miss something important? What are the consequences to not seeing that!?" It matters not if the notification was unimportant;people //hate// the feeling of "missing out", and the current low duration induces those feelings.

This patch increases the default time on-screen to 7 seconds, greatly increasing the liklihood that the user will actually notice and read the notification before it disappears. It also accordingly increases the `short` and `long` durations too.

Test Plan:
Test in Discover:
- Install Flatpak backend
- Go to {nav Settings} page
- Click hamburger menu for Flatpah backend
- Choose "Add Source"
- Enter "aaaaaaaaaaaaaa" in the text field and click {nav OK}

With this patch, the message that appears lasts for 7 seconds, and you can actually read it in time before it disappears.

Reviewers: #kirigami, mart

Reviewed By: #kirigami, mart

Subscribers: plasma-devel, apol

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D11009