Allow OverlaySheet clients to omit the built-in close button
Summary: Most of the time we want an OverlaySheet to draw a close button so its clients don't have to. But sometimes clients want to draw their own close buttons, for example to use the OverlaySheet as a dialog box with a {nav Cancel} button (which we do in Discover). In these circumstances, we don't need OverlaySheet to draw its own close button; this patch makes that possible.

Test Plan:
Tested in Discover after applying the patch and adding `showCloseButton: false` to DiscoverWindow.qml's proceedDialog:
{F5747561}

Reviewers: mart, #kirigami

Reviewed By: mart, #kirigami

Subscribers: plasma-devel, apol

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D11197