checkable FormLayout items
Summary:
In some cases, form items go together with a checkbox which can be used to disable the item.
An example for such an item is the line edit in the first row of the screenlocker kcm.
Here, a checkbox is prepended to the line edit's label.
{F5752567}

In Kirigami, there is (afaik) no way to add a widget to a FormLayout item's buddyComponent.
This patch adds an optional checkbox to the buddyComponent.

A place where I would like to use this is the fonts kcm: https://phabricator.kde.org/T7927#132277

Test Plan:
I added a checkable item to the FormLayout gallery page.

I also rewrote the fonts kcm (not comitted yet), which now looks like this:
{F5758858}

Reviewers: #vdg, #kirigami, mart

Reviewed By: #vdg, #kirigami, mart

Subscribers: mart, ngraham, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D11289