watch the tabletMode from KWin
Summary:
Add api in the C++ library that watches the tablet mode property exposed from
KWin. use dbus, but only on linux,bsd etc, not on Android, Windows etc.
connect it to Settings.isMobile which was already used for it (even tough
it introduces a naming inconsistence, but needed for retrocompatibility)

Test Plan: tested to switch on the fly on Wayland with the kwin tablet mode sensor working

Reviewers: #plasma, #kirigami, davidedmundson

Reviewed By: #plasma, davidedmundson

Subscribers: ngraham, plasma-devel

Tags: #plasma, #kirigami

Differential Revision: https://phabricator.kde.org/D10112