Drag handle to reorder items in a ListView
Summary:
A component to be inserted in list items of any kind, when used in ListViews
when the user wishes to be able to reorder items by drag.

The handle must be used in a list item which is not directly
the view's delegate, but its main child. the main delegate will act
as empty space placeholder.
It is recomended to use the DelegateRecycler component for this purpose

Test Plan: used in the list view gallery page

Reviewers: #kirigami, #vdg, hein

Subscribers: alex-l, abetts, rkflx, ngraham, zzag, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D12960