Collapsible Sidebar Mode
Summary:
This adds a new mode for OverlayDrawer, mostly for GlobalDrawer but applications can implement their own collapsing visuals in an OverlayDrawer. It is a Desktop-only feature.
NOTE: this is explicitly and only for drawers: it is NOT intended for content which semantically should be in a Page in the main PageRow: if needed there similar such behavior should be implemented for the Page and the PageRow's defaultColumnWidth instead.

a new property called "collapsible" controls wether the drawer supports the collapsed mode.
NOTE: It is not supported neither in mobile mode nor on modal, overlay drawers

the property "collapsed" controls whether the drawer is collapsed or not
NOTE: it's supported only in "collapsible" mode

Test Plan: tested in the gallery app

Reviewers: #kirigami

Subscribers: ngraham, plasma-devel

Tags: #kirigami, #vdg

Differential Revision: https://phabricator.kde.org/D14211