Fix the Kirigami QRC static plugin generation
Summary:
This patch moves the qrc from "raw" to generated using CMake configure file and add 4 missing files.

Since some months, a few files were missing from the QRC. It worked fine on system where Kirigami is installed (and recent) because it would load them from the file system instead of static resources. However it didn't work so well on systems where an older version (or no version) are installed.

Because ActionMenuItem is has different versions depending on the Qt version, the qrc had to be dynamic. The use of absolute path was also necessary because `rcc` could no longer find the files because it uses the relative path from the `.qrc` absolute path.

Test Plan: Not much honestly, it works on my system, other might want to double check.

Reviewers: lepagevalleeemmanuel

Subscribers: bearsh, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D14673