Explicitly request Qt 5.7's QtQuick to make use of Connections.enabled
Qt 5 now has stricter checking on the use of QML properties, and only
exposes properties introduced in a minor release if you explicitly
request QtQuick from that minor release or later.

This causes some Kirigami users to fail to launch due to use of the
"enabled" property in the Connections QML component, since the Kirigami
QML code requested QtQuick 2.5 (from Qt 5.5) but that property was
introduced only in Qt 5.7. Fixed by requesting the appropriate QtQuick
version.

All 2 uses of "enabled" are fixed here.

Differential Revision: https://phabricator.kde.org/D14984