Allow creation of separator Actions
The QAction class allows an action to be considered a separator. In Kirigami Action the separator property has been added to serve the same purpose. A set of delegates have been changed to draw a separator in case that an Action with separator:true has been found in the model.

Differential Revision: https://phabricator.kde.org/D15143