Fix static Kirigami (about page etc)
Without this patch, if you tried to use the AboutPage component
in a static Kirigami app, the QML engine would complain that the
file was missing (of course). Further, once that was fixed, the
function i18n also does not exist unless you add it, and since
everything else in Kirigami uses qsTr, it seems to make sense to
use that in AboutPage and UrlButton as well.

Differential Revision: https://phabricator.kde.org/D17961