Add an ActionTextField component
Summary:
<old>This is based on the searchField from Discover. It's a suggestion from @ngraham and I find it's a great idea.

Note: Because it's based on the searchField from Discover licensed under GPL2-or-later and Kirigami is licensed under LGPL2-or-later, this patch need the authorization from @apol and @ngraham, before landing.</old>

Now it's a generic TextField with the possibility to add buttons left and a button right. Also, update the Sidebar example to use this component:

{F6602619} {F6602620}

Test Plan: For now kirigami build. I wasn't only able to test this component by copying this component to the example directory.

Reviewers: apol, #kirigami, ngraham, mart

Reviewed By: #kirigami, mart

Subscribers: broulik, mart, apol, ngraham, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D18716