make handle follow arbitrary height of the drawer
Summary:
If the height of the drawer is set to a different value from the applicationWindow height, then the drawer handle is placed at the bottom of the drawer, following the new arbitrary height value.

with height property unset:

{F7076769}

{F7076768}

with the height property set to  an arbitrary value:
{F7076774}

{F7076773}

use case:
when there is a Page with a footer and it is not inside the pageStack, the handle drawer covers the footer, by setting now the drawer height to application window height minus the height of the footers the content is now not overlapping. This allows placing the drawer and handle more freely.

before:
{F7076806}

after:
{F7076805}

{F7076804}

Reviewers: mart, #kirigami

Reviewed By: mart, #kirigami

Subscribers: camiloh, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D22648