Allow usage of QQC2 actions on Kirigami components and now make K.Action based on QQC2.Action
Summary: The ideal is to be able to add regular qqc2 actions to the actiontoolbar, pageheaders, globaldrawers and different components, and make kirigami.action based on qqc2.action, so implementations are not duplicated.

Test Plan:
So far i have tested this with different kirigami components, such as pageheadertoolbar, globaldrawers, kirigamiactiontoolbar, and it is working.
There is an issue now with the PrivateActionToolButton, since K.Actions is based on QQC2.Action, it make sthe control to draw two icons, one from PrivateActionToolBar own contentItem implementation and another one from the style background

I will continue testing this, and pls let me know how i could proceed better with the PrivateActionToolButton issue.

Reviewers: #kirigami, mart

Reviewed By: #kirigami, mart

Subscribers: astippich, apol, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D22974