Rework the visibleActions logic in SwipeListItem
Summary:
visibleActions is really only used to check whether there are
any actions visible at all, the number is irrelevant. For some
unknown reason the old logic would occasionally get caught in
some out of order silliness (so we'd end up with a negative
number of visible actions), but with this logic we only check
whether there is any visible actions at all, so the whole thing
is less stateful.

Test Plan:
Using Peruse on mobile, load up the Store and see that with the
old logic there are no handles on the items. With the new
logic there is.

Reviewers: #kirigami, mart

Reviewed By: #kirigami, mart

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D23109