Replace the custom overflow button with PrivateActionToolButton in ActionToolBar
Summary:
With the fixes done to PrivateActionToolBar it now behaves a lot nicer
and we should try to make sure all other places that use a similar
button style use the same component.

Test Plan:
Start Kirigami Gallery, go to the Cards Layout page, the overflow buttons there
should now properly open and close the menus and not get stuck in a checked
state.

Reviewers: #kirigami, mart

Reviewed By: #kirigami, mart

Subscribers: apol, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D23502