Be more explicit about action property types
Summary:
Now that Kirigami.Action is a Controls.Action, we can be more explicit
about using the Action type where we store it as a property. This adds a
bit of type safety and additionally allows binding of the action's
properties in more cases.

Test Plan: Run kirigami2gallery, see that things that use actions still work correctly.

Reviewers: #kirigami, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D24500