Don't show context drawer if all actions are invisible
Summary: Discover can be in a state where the only action in the context drawer is invisible. In that case an expty drawer just with the work "Actions" is shown, which isn't pretty.

Test Plan: Tested with affected discover page and unaffected KDE Connect page in mobile mode

Reviewers: mart

Reviewed By: mart

Subscribers: ngraham, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D24648