Ensure that GlobalDrawer topContent always stays on top
Summary:
Right now, if you define `topContent` for your GlobalDrawer, it's added to the scrollview
so it can actually scroll out of sight, making it not always on top.

This patch moves the `topContent` above the scrollview so it always stays on top.

Test Plan:
Discover's GlobalDrawer toolbar no longer strangely scrolls out of view: {F7680325}

No regressions in Discover when the view is not scrollable

No regressions that I could detect in Kirigami Gallery, though this probably needs lots
more testing to ensure that nothing has exploded

BUG: 389533
FIXED-IN: 5.64

Reviewers: #kirigami, mart, apol, ahiemstra

Reviewed By: ahiemstra

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D25019