FormLayout: Fix label height if wide mode is false
Summary:
previously the height of the label was based on the content's height. However this doesn't work in non wide mode. With this change it preserves the previous behaviour in wide mode were it makes sense but uses a constant height in non-wide mode.

Before:
{F7744926}

After:
{F7744929}

Test Plan: FormLayouts with large content items don't have large margins between the items anymore when wideMode = false.

Reviewers: #kirigami, #plasma:_mobile, mart

Reviewed By: #kirigami, #plasma:_mobile, mart

Subscribers: mart, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D25175