Switch ToolBarPageHeader to use the icon collapse behaviour from ActionToolBar
Summary:
This removes the icon collapse behaviour of ToolBarPageHeader in favour
of the implementation that is now implemented in ActionToolBar. It also
fixes a small performance regression that was introduced with the action
overflow behaviour.

Depends on D25177

Test Plan:
Kirigami Gallery still shows actions, the main/left/right actions change to
icons at the appropriate moments.

Reviewers: #kirigami, mart

Reviewed By: #kirigami, mart

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D25178