Add a displayHint property to Action
Summary:
This property can be used to indicate preferences on how to display the
action to implementations using Action. It mimics some of the behaviour
of QAction's priority property, but is more explicit and allows some
behaviour the priority property would not.

Additionally, this changes PrivateActionToolButton to make use of the
displayHint property for its showText and showMenuArrow properties.

Test Plan:
Tool buttons created through actions in Kirigami gallery still show
their text and menu arrows.

Reviewers: #kirigami, mart

Reviewed By: #kirigami, mart

Subscribers: broulik, leinir, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D25176