ActionToolBar: Automatically change to icon-only for actions marked KeepVisible
Summary:
This changes the layouting code in ActionToolBar to support reducing
actions to IconOnly before hiding them for actions that have a
displayHint set to KeepVisible. This mimics behaviour that is currently
implemented in ToolBarHeader, where the main, left and right actions
will be reduced to an icon-only version when the page gets below a
certain width. The main difference in behaviour is that instead of
switching all three actions at the same time, the actions will be
switched one by one, which I feel makes more sense when using
ActionToolBar standalone.

Depends on D25176

Test Plan:
With Kirigami Gallery, on the misc widgets page, behaviour of the second toolbar
is still the same.

Reviewers: #kirigami, mart

Reviewed By: #kirigami, mart

Subscribers: mart, broulik, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D25177