header property for the drawer
Summary:
GlobalDrawer has the new property header, which does not scroll.
It can be used for toolbars that need to match with the global kirigami toolbar.
the banner image by default is *not* shown on desktop but only on mobile
the result makes easier to make apps look more at home on desktops

Test Plan: apps that don't use it still look the same

Reviewers: #kirigami

Subscribers: ngraham, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D25425