Document Kirigami.Theme Better
Summary:
Working on a tertiary project I ran into how Kirigami.Theme has depreciated colors that aren't documented as such on the apidox. The current discoverability of this is that a user would browse the docs to Theme from the index, start using depreciated colors, get warnings, and only then discover that Kirigami.Theme is attached - probably via [[ https://notmart.org/blog/2017/10/kirigami-and-color-palettes/ | this blog post ]] by @mart.

This adds a blurb describing how the property is attached and one should refer to and use PlatformTheme for contemporary color management.

There might be a bug here, though -"Remarks: This component is a singleton" is duplicated when I build this locally and it runs on to the first paragraph. Not sure if its just a quirk in my apidox builder setup.

Reviewers: apol, davidedmundson

Reviewed By: apol, davidedmundson

Subscribers: davidedmundson, davidre, apol, mart, plasma-devel

Tags: #documentation, #kirigami

Differential Revision: https://phabricator.kde.org/D25227