Account for a mnemonic label to go back to ""
Summary:
We are discarding the value if there's no weights which meant that when
setting a button to "", the text provided by mnemonicLabel wouldn't update.
This patch changes it so it only skips the weights computation if there's no
text.

Test Plan:
Tested with patch that removes the text of some buttons when the
screen is smaller on Discover.

Reviewers: #kirigami, mart, ngraham

Reviewed By: ngraham

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D25618