Remove header top margin from private ScrollView
Summary:
With the scrollbars taking up actual width, the top margin of the vertical
scrollbar being set to the list header looks really awkward and buggy. This
removes the top margin, making the scrollbar take the entire flickable's
height. This change also applies to overlay scrollbars as I consider the
margin behaviour to be incorrect even for those, as the top margin will
change based on the header height, which causes the scrollbar to resize
when the header shrinks.

Test Plan:
Before:

{F8022825}

After:

{F8022827}

Reviewers: #kirigami, ngraham, mart

Reviewed By: #kirigami, ngraham, mart

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D27029