Replace qmltestrunner with a custom executable that effectively does the same
Summary:
CI is currently failing due to LD_PRELOAD of the ASAN libraries failing for some
reason. Since it is not clear what causes this problem, instead create a custom
executable that does the same as qmltestrunner but will be compiled using ASAN
like the rest of the library.

Test Plan:
Tests still pass. I used the same solution for kquickcharts where it has been working
for a while.

Reviewers: #kirigami, apol

Reviewed By: apol

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D27699