Fallback to a simple rectangle when using software rendering
Summary:
If the scene graph is using software rendering, the custom shaders used for
ShadowedRectangle will not work. So in that case, create a simple rectangle
node and use that. This loses a bunch of features of the item but at least
shows something.

Test Plan: ShadowedRectangleTest with QT_QUICK_BACKEND=software set renders a rectangle.

Reviewers: #kirigami, mart

Reviewed By: #kirigami, mart

Subscribers: plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D28022