Add ShadowedImage and ShadowedTexture items
Summary:
This adds the ShadowedImage and ShadowedTexture items. The first is a
QML-only item that makes use of the second. The second is a
ShadowedRectangle that uses a texture as color source for the item
instead of just a color.

It also does some cleanup on the scenegraph/shader related bits
introduced with ShadowedRectangle. Most importantly, it removes the
mostly duplicated core shaders in favour of a simple solution using
a #define.

Test Plan:
There's a manual test included. It should look like this:

{F8195115}

Reviewers: #kirigami, mart

Reviewed By: #kirigami, mart

Subscribers: mart, plasma-devel

Tags: #kirigami

Differential Revision: https://phabricator.kde.org/D28222